Move once-only components into layout directory
