+++ src/layout/index.ts
@@ -1,0 +1,5 @@
+import Jumbo from "./jumbo/Jumbo.vue";
+import Navigation from "./navigation/Navigation.vue";
+import FooterBar from "./FooterBar.vue";
+
+export { Jumbo, Navigation, FooterBar };
